requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
bitarray==2.9.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
cryptography==43.0.1
cytoolz==0.12.3
Django==5.1.1
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.3
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
frozenlist==1.4.1
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
h11==0.14.0
hexbytes==1.2.1
httplib2==0.22.0
idna==3.8
injector==0.22.0
mnemonic==0.21
multidict==6.1.0
parsimonious==0.10.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
pyparsing==3.1.4
python-decouple==3.8
pyunormalize==15.1.0
pywin32==306
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
rsa==4.9
sqlparse==0.5.1
toolz==0.12.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
web3==7.2.0
websockets==13.0.1
yarl==1.11.1
